backend/requirements.txt
Django>=4.2,<5.0
channels>=4.2
daphne>=4.1,<5.0
djangorestframework>=3.15
whitenoise>=6.9,<7.0
django-cors-headers>=4.0
djangorestframework-simplejwt>=2.1
django-extensions>=3.2,<4.2
django-jazzmin>=3.0,<3.1
gunicorn>=22.1.0
Pillow>=10.0,<11.0
numpy>=1.26.4
ftfy
regex
sentencepiece
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.41.0
pytest>=8.3,<9.0
pytest-django>=4.8,<5.0
pytest-mock>=3.14,<4.0